pyproject.toml
[project]
name = "word2vec"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.6.1",
    "huggingface-hub>=1.6.0",
    "hydra-core>=1.3.2",
    "loguru>=0.7.3",
    "numpy>=2.4.2",
    "scipy>=1.17.1",
]

[tool.ty.rules]
all = "warn"
